Change how we ensure permissions for tox files are correct
Prior to this commit we ensured we didn't create root files in the /app
directory by using the --user flag to run.

This caused problems because our pre-built docker image didn't necessarily have
a user with that id in passwd.

This commit instead aims to discover any files in the /app directory with root
permissions that were created by tox while running as root.
We then reset those files specifically to the uid and gid set with the BE_UID
and BE_GID env variables.

It's hacky as hell but it's the best we could think of for now.

An alternative might have been to copy the entire /app directory when we docker
run and then work on the copy.
The disadvantage of this is that we can't know which files we want to copy back
to the real /app after running tox.